worker_filter_score handlers arq

+++ worker_filter_score/src/messages/message_score_avg.rs
@@ -3,5 +3,5 @@ use serde::Deserialize;
 #[derive(Deserialize)]
 pub struct MessageScoreAvg {
     pub opcode: u8,
-    pub payload: Option<i32>
+    pub payload: Option<f32>
 }
